# api/openapi.yaml
openapi: 3.0.3
info:
  title: Otedama Mining Pool API
  description: |
    Professional P2P Mining Pool & DEX Platform API with BTC-only payouts.
    
    ## Features
    - RESTful API for pool statistics and management
    - WebSocket API for real-time mining operations
    - Comprehensive authentication and security
    - Support for 13 cryptocurrencies and 10 mining algorithms
    
    ## Authentication
    All authenticated endpoints require either:
    - Bearer token (JWT) in Authorization header
    - API key in X-API-Key header
  version: 0.7.0
  contact:
    name: Otedama Support
    url: https://github.com/shizukutanaka/Otedama
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3333
    description: Local development server
  - url: http://localhost:8080
    description: Local API server
...
paths:
...
  /api/prices:
    get:
      tags:
        - Public
      summary: Get cryptocurrency prices
      description: Get current cryptocurrency prices in BTC and USD
      operationId: getPrices
      responses:
        '200':
          description: Current prices
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Prices'
        '429':
          $ref: '#/components/responses/RateLimitExceeded'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
...
  /api/dex/pools:
    get:
      tags:
        - DEX
      summary: List liquidity pools
      description: Get all available liquidity pools
      operationId: getLiquidityPools
      parameters:
        - name: token0
          in: query
          description: Filter by first token
          schema:
            type: string
        - name: token1
          in: query
          description: Filter by second token
          schema:
            type: string
      responses:
        '200':
          description: List of liquidity pools
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/LiquidityPool'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from authentication
    apiKey:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key for programmatic access
  schemas:
...
    Prices:
      type: object
      additionalProperties:
        type: object
        properties:
          usd:
            type: number
            description: Price in USD
          btc:
            type: number
            description: Price in BTC
          change24h:
            type: number
            description: 24-hour change percentage
...
    LiquidityPool:
      type: object
      properties:
        id:
          type: string
        token0:
          type: string
        token1:
          type: string
        reserve0:
          type: number
        reserve1:
          type: number
        totalShares:
          type: number
        fee:
          type: number
          description: Trading fee (0.003 = 0.3%)
        volume24h:
          type: number
        tvl:
          type: number
          description: Total value locked in USD
...
    Error:
      type: object
      properties:
        error:
          type: string
          description: Error message
        code:
          type: string
          description: Error code
        details:
          type: object
          description: Additional error details
  responses:
...
    RateLimitExceeded:
      description: Rate limit exceeded
      headers:
        X-RateLimit-Limit:
          schema:
            type: integer
          description: Request limit per window
        X-RateLimit-Remaining:
          schema:
            type: integer
          description: Remaining requests in window
        X-RateLimit-Reset:
          schema:
            type: integer
          description: Time when the rate limit resets (Unix timestamp)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: Rate limit exceeded
            code: RATE_LIMIT_EXCEEDED
    ServiceUnavailable:
      description: Service unavailable
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: Service temporarily unavailable
            code: SERVICE_UNAVAILABLE